+++ src/main/java/com/github/jlangch/venice/util/excel/ExcelSheetWriter.java
@@ -157,6 +157,11 @@ public class ExcelSheetWriter<T> {
         return this;
     }
 
+    public ExcelSheetWriter<T> style(final int row1, final int col1, final String stylename) {
+        sheet.setStyle(row1-1, col1-1, stylename);
+        return this;
+    }
+
     public ExcelSheetWriter<T> bgColor(final int row1, final int col1, final Color bgColor) {
         sheet.setBgColor(row1-1, col1-1, bgColor);
         return this;