+++ ChangeLog.md
@@ -14,7 +14,7 @@ and this project adheres to
 
 ### Added 
 
-- ...
+- improvements `:excel` module cell styling